+++ texlive/texk/lcdf-typetools/lcdf-typetools-src/libefont/t1item.cc
@@ -2,7 +2,7 @@
 
 /* t1item.{cc,hh} -- items in a Type 1 font
  *
- * Copyright (c) 1998-2016 Eddie Kohler
+ * Copyright (c) 1998-2018 Eddie Kohler
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the Free
@@ -731,12 +731,12 @@ Type1Subr::gen(Type1Writer &w)
         for (int i = 0; i < w.lenIV(); i++) {
             unsigned char c = (unsigned char)(r >> 8);
             *t++ = c;
-            r = ((c + r) * t1C1 + t1C2) & 0xFFFF;
+            r = ((c + r) * (uint32_t) t1C1 + t1C2) & 0xFFFF;
         }
         for (int i = 0; i < len; i++, data++) {
             unsigned char c = (*data ^ (r >> 8));
             *t++ = c;
-            r = ((c + r) * t1C1 + t1C2) & 0xFFFF;
+            r = ((c + r) * (uint32_t) t1C1 + t1C2) & 0xFFFF;
         }
 
         w.print((char *)buf, len + w.lenIV());
